Revert "biometric design changes"
