--- app/res/drawable/grey_small_button.xml
@@ -1,18 +1,0 @@
-<vector
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:aapt="http://schemas.android.com/aapt"
-    android:width="135dp"
-    android:height="40dp"
-    android:viewportWidth="135"
-    android:viewportHeight="40"
-    >
-    <group>
-        <clip-path
-            android:pathData="M20 0H115C126.046 0 135 8.95431 135 20C135 31.0457 126.046 40 115 40H20C8.95431 40 0 31.0457 0 20C0 8.95431 8.95431 0 20 0Z"
-            />
-        <path
-            android:pathData="M0 0V40H135V0"
-            android:fillColor="#D9D9D9"
-            />
-    </group>
-</vector>
