--- app/res/drawable/rounded_rect.xml
@@ -1,13 +1,0 @@
-<!-- res/drawable/rounded_background.xml -->
-<shape xmlns:android="http://schemas.android.com/apk/res/android">
-    <solid android:color="#09666666"/> <!-- Background color, white in this case -->
-    <corners android:radius="16dp"/> <!-- Adjust the radius as needed -->
-    <padding
-        android:left="16dp"
-        android:top="16dp"
-        android:right="16dp"
-        android:bottom="16dp"/>
-    <stroke
-        android:width="1dp"
-        android:color="#09666666"/> <!-- Border color -->
-</shape>
