--- app/res/layout/activity_biometric_signup_page.xml
@@ -1,167 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:padding="16dp"
-    tools:context="org.commcare.activities.connect.BiometricSignupPage">
-
-
-    <ImageView
-        android:id="@+id/ivCommCareLogo"
-        android:layout_width="250dp"
-        android:layout_height="70dp"
-        android:layout_marginTop="59dp"
-        android:layout_marginBottom="43dp"
-        android:src="@drawable/commcare_diamgi_logo" />
-
-    <TextView
-        android:id="@+id/text_view_name"
-        android:layout_width="wrap_content"
-        android:layout_height="23dp"
-        android:layout_below="@id/ivCommCareLogo"
-        android:layout_marginTop="2dp"
-        android:layout_marginBottom="8dp"
-        android:text="@string/app_lock"
-        android:textColor="@color/black"
-        android:textSize="18sp"
-        android:textStyle="bold" />
-
-    <!-- Unlock with Biometric Section -->
-    <LinearLayout
-        android:id="@+id/biometricSection"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_below="@id/text_view_name"
-        android:layout_marginTop="5dp"
-        android:background="@drawable/rounded_rect"
-        android:orientation="vertical"
-        android:padding="16dp">
-
-        <TextView
-            android:id="@+id/tvUnlockWithBiometric"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginBottom="8dp"
-            android:text="@string/unlock_with_biometric"
-            android:textSize="16sp"
-            android:textStyle="bold" />
-
-        <TextView
-            android:id="@+id/tvBiometricDescription"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginBottom="16dp"
-            android:lines="3"
-            android:text="@string/unlock_with_biometric_text"
-            android:textSize="16sp" />
-
-        <LinearLayout
-            android:layout_width="wrap_content"
-            android:layout_height="40dp"
-            android:layout_marginBottom="16dp"
-            android:orientation="horizontal">
-
-            <!-- Fingerprint Icon -->
-            <ImageView
-                android:id="@+id/ivFingerprint"
-                android:layout_width="40dp"
-                android:layout_height="40dp"
-                android:layout_marginEnd="16dp"
-                android:src="@drawable/fingerprint"
-                app:tint="@color/pattern_color"
-                />
-
-            <!-- Face Icon -->
-            <ImageView
-                android:id="@+id/ivFace"
-                android:layout_width="40dp"
-                android:layout_height="40dp"
-                android:layout_marginEnd="16dp"
-                android:src="@drawable/pattern"
-                app:tint="@color/pattern_color" />
-
-            <!-- Pattern Icon -->
-            <ImageView
-                android:id="@+id/ivPattern"
-                android:layout_width="40dp"
-                android:layout_height="40dp"
-                android:layout_marginEnd="16dp"
-                android:src="@drawable/dialpad"
-                app:tint="@color/pattern_color"/>
-
-        </LinearLayout>
-
-        <Switch
-            android:id="@+id/switchBiometric"
-            android:layout_width="wrap_content"
-            android:layout_height="48dp"
-            android:layout_gravity="right"
-            />
-
-    </LinearLayout>
-
-    <!-- Setup 6 Digit PIN Section -->
-    <LinearLayout
-        android:id="@+id/pinSection"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_below="@id/biometricSection"
-        android:layout_marginTop="32dp"
-        android:gravity="center_vertical"
-        android:orientation="horizontal">
-
-        <ImageView
-            android:id="@+id/ivLockIcon"
-            android:layout_width="40dp"
-            android:layout_height="40dp"
-            android:layout_marginEnd="16dp"
-            android:src="@drawable/lock" />
-
-        <TextView
-            android:id="@+id/tvSetupPin"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="@string/set_up_pin"
-            android:textColor="@color/black"
-            android:textSize="16sp" />
-
-        <LinearLayout
-            android:id="@+id/btnContinue"
-            android:layout_width="135dp"
-            android:layout_height="52dp"
-            android:layout_below="@id/pinSection"
-            android:layout_alignParentEnd="true"
-            android:layout_marginStart="60dp"
-            android:background="@drawable/grey_small_button"
-            android:clickable="true"
-            android:focusable="true"
-            android:gravity="right"
-            android:orientation="horizontal"
-            android:padding="12dp">
-
-            <TextView
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_gravity="center"
-                android:layout_marginEnd="16dp"
-                android:text="@string/connect_verify_button_configured"
-                android:textAllCaps="false"
-                android:textColor="@color/black"
-                android:textSize="16sp" />
-
-            <ImageView
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_gravity="center_vertical"
-                android:adjustViewBounds="true"
-                android:maxWidth="18dp"
-                android:maxHeight="18dp"
-                android:scaleType="fitCenter"
-                android:src="@drawable/arow_icon" />
-        </LinearLayout>
-    </LinearLayout>
-
-
-</RelativeLayout>
