+++ app/res/values/colors.xml
@@ -146,6 +146,4 @@
     <!-- Button colors -->
     <color name="button_primary">#005ab2</color>
     <color name="button_disabled">#d0e3ff</color>
-
-    <color name="pattern_color">#1c1b1f</color>
 </resources>
