+++ app/src/org/commcare/connect/ConnectTask.java
@@ -30,8 +30,8 @@ public enum ConnectTask {
             ConnectIdRegistrationActivity.class),
     CONNECT_REGISTRATION_CONFIGURE_BIOMETRICS(ConnectConstants.ConnectIdTaskIdOffset + 5,
             ConnectIdBiometricConfigActivity.class),
-//    CONNECT_REGISTRATION_UNLOCK_BIOMETRIC(ConnectConstants.ConnectIdTaskIdOffset + 6,
-//            ConnectIdBiometricUnlockActivity.class),
+    CONNECT_REGISTRATION_UNLOCK_BIOMETRIC(ConnectConstants.ConnectIdTaskIdOffset + 6,
+            ConnectIdBiometricUnlockActivity.class),
     CONNECT_REGISTRATION_VERIFY_PRIMARY_PHONE(ConnectConstants.ConnectIdTaskIdOffset + 7,
             ConnectIdPhoneVerificationActivity.class),
     CONNECT_REGISTRATION_CHANGE_PRIMARY_PHONE(ConnectConstants.ConnectIdTaskIdOffset + 8,
@@ -72,8 +72,8 @@ public enum ConnectTask {
             ConnectIdPinActivity.class),
     CONNECT_RECOVERY_CONFIGURE_BIOMETRICS(ConnectConstants.ConnectIdTaskIdOffset + 28,
             ConnectIdBiometricConfigActivity.class),
-//    CONNECT_RECOVERY_UNLOCK_BIOMETRIC(ConnectConstants.ConnectIdTaskIdOffset + 29,
-//            ConnectIdBiometricUnlockActivity.class),
+    CONNECT_RECOVERY_UNLOCK_BIOMETRIC(ConnectConstants.ConnectIdTaskIdOffset + 29,
+            ConnectIdBiometricUnlockActivity.class),
     CONNECT_VERIFY_ALT_PHONE_MESSAGE(ConnectConstants.ConnectIdTaskIdOffset + 30,
             ConnectIdMessageActivity.class),
     CONNECT_VERIFY_ALT_PHONE(ConnectConstants.ConnectIdTaskIdOffset + 31,
